x11.hidd: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@43585 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/all-hosted/hidd/x11/makexkeytable.c
@@ -343,7 +343,7 @@ void doaskuser(char *title,int rawkeystart, int numkeys, KeySym k)
 		    unsigned char kc = (unsigned char)((XKeyEvent *)&event)->keycode;
 		    table[kc] = rawkeystart++;
 		    
-		    sprintf(s, "%03ld",(int)((XKeyEvent *)&event)->keycode);
+		    sprintf(s, "%03ld",(long)((XKeyEvent *)&event)->keycode);
 		    XDrawString(dpy, w, gc, 50, 20, s, strlen(s));
 
 		} else {
